code restructuring

+++ server/Downlink_channel.cpp
@@ -6,7 +6,7 @@
 #include <sys/socket.h>
 #include <netinet/in.h>
 #include <sys/epoll.h>
-#include <errno.h>
+#include <cerrno>
 #include <iostream>
 #include <string>
 #include <cstring>
@@ -20,7 +20,7 @@ Downlink_channel::Downlink_channel(int port)
     this->socket_fd = socket(AF_INET, SOCK_DGRAM, 0);
 
     this->addr.sin_family = AF_INET;
-    this->addr.sin_port = htons(this->port);
+    this->addr.sin_port = htons(static_cast<uint16_t>(this->port));
     this->addr.sin_addr.s_addr = htons(INADDR_ANY);
 
     if(bind(this->socket_fd, (struct sockaddr*)&addr, sizeof(addr)) == -1)
